+++ core/src/scheduler/resource/CpuResource.cpp
@@ -24,7 +24,7 @@ namespace scheduler {
 
 std::ostream&
 operator<<(std::ostream& out, const CpuResource& resource) {
-    out << resource.Dump();
+    out << resource.Dump().dump();
     return out;
 }