+++ tests/milvus_python_test/test_index.py
@@ -8,14 +8,15 @@ import pdb
 import threading
 from multiprocessing import Pool, Process
 import numpy
+import sklearn.preprocessing
 from milvus import Milvus, IndexType, MetricType
 from utils import *
 
-nb = 100000
+nb = 10000
 dim = 128
 index_file_size = 10
 vectors = gen_vectors(nb, dim)
-vectors /= numpy.linalg.norm(vectors)
+vectors = sklearn.preprocessing.normalize(vectors, axis=1, norm='l2')
 vectors = vectors.tolist()
 BUILD_TIMEOUT = 60
 nprobe = 1
@@ -36,8 +37,11 @@ class TestIndexBase:
         scope="function",
         params=gen_simple_index_params()
     )
-    def get_simple_index_params(self, request):
-        yield request.param
+    def get_simple_index_params(self, request, args):
+        if "internal" not in args:
+            if request.param["index_type"] == IndexType.IVF_SQ8H:
+                pytest.skip("sq8h not support in open source")
+        return request.param
 
     """
     ******************************************************************
@@ -65,8 +69,10 @@ class TestIndexBase:
         method: create table and add vectors in it, check if added successfully
         expected: raise exception
         '''
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
         with pytest.raises(Exception) as e:
-            status = dis_connect.create_index(table, random.choice(gen_index_params()))
+            status = dis_connect.create_index(table, index_param)
 
     @pytest.mark.timeout(BUILD_TIMEOUT)
     def test_create_index_search_with_query_vectors(self, connect, table, get_index_params):
@@ -179,12 +185,14 @@ class TestIndexBase:
     def test_create_index_table_not_existed(self, connect):
         '''
         target: test create index interface when table name not existed
-        method: create table and add vectors in it, create index with an random table_name
+        method: create table and add vectors in it, create index
             , make sure the table name not in index
         expected: return code not equals to 0, create index failed
         '''
         table_name = gen_unique_str(self.__class__.__name__)
-        status = connect.create_index(table_name, random.choice(gen_index_params()))
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
+        status = connect.create_index(table_name, index_param)
         assert not status.OK()
 
     def test_create_index_table_None(self, connect):
@@ -194,8 +202,10 @@ class TestIndexBase:
         expected: return code not equals to 0, create index failed
         '''
         table_name = None
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
         with pytest.raises(Exception) as e:
-            status = connect.create_index(table_name, random.choice(gen_index_params()))
+            status = connect.create_index(table_name, index_param)
 
     def test_create_index_no_vectors(self, connect, table):
         '''
@@ -203,32 +213,34 @@ class TestIndexBase:
         method: create table and add no vectors in it, and then create index
         expected: return code equals to 0
         '''
-        status = connect.create_index(table, random.choice(gen_index_params()))
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
+        status = connect.create_index(table, index_param)
         assert status.OK()
 
     @pytest.mark.timeout(BUILD_TIMEOUT)
-    def test_create_index_no_vectors_then_add_vectors(self, connect, table):
+    def test_create_index_no_vectors_then_add_vectors(self, connect, table, get_simple_index_params):
         '''
         target: test create index interface when there is no vectors in table, and does not affect the subsequent process
         method: create table and add no vectors in it, and then create index, add vectors in it
         expected: return code equals to 0
         '''
-        status = connect.create_index(table, random.choice(gen_index_params()))
+        index_param = get_simple_index_params
+        status = connect.create_index(table, index_param)
         status, ids = connect.add_vectors(table, vectors)
         assert status.OK()
 
     @pytest.mark.timeout(BUILD_TIMEOUT)
-    def test_create_same_index_repeatedly(self, connect, table):
+    def test_create_same_index_repeatedly(self, connect, table, get_simple_index_params):
         '''
         target: check if index can be created repeatedly, with the same create_index params
         method: create index after index have been built
         expected: return code success, and search ok
         '''
         status, ids = connect.add_vectors(table, vectors)
-        index_params = random.choice(gen_index_params())
-        # index_params = get_index_params
-        status = connect.create_index(table, index_params)
-        status = connect.create_index(table, index_params)
+        index_param = get_simple_index_params
+        status = connect.create_index(table, index_param)
+        status = connect.create_index(table, index_param)
         assert status.OK()
         query_vec = [vectors[0]]
         top_k = 1
@@ -243,16 +255,19 @@ class TestIndexBase:
         method: create another index with different index_params after index have been built
         expected: return code 0, and describe index result equals with the second index params
         '''
+        nlist = 16384
         status, ids = connect.add_vectors(table, vectors)
-        index_params = random.sample(gen_index_params(), 2)
+        index_type_1 = IndexType.IVF_SQ8
+        index_type_2 = IndexType.IVFLAT
+        index_params = [{"index_type": index_type_1, "nlist": nlist}, {"index_type": index_type_2, "nlist": nlist}]
         logging.getLogger().info(index_params)
-        status = connect.create_index(table, index_params[0])
-        status = connect.create_index(table, index_params[1])
-        assert status.OK()
+        for index_param in index_params:
+            status = connect.create_index(table, index_param)
+            assert status.OK()
         status, result = connect.describe_index(table)
-        assert result._nlist == index_params[1]["nlist"]
+        assert result._nlist == nlist
         assert result._table_name == table
-        assert result._index_type == index_params[1]["index_type"]
+        assert result._index_type == index_type_2
 
     """
     ******************************************************************
@@ -286,7 +301,7 @@ class TestIndexBase:
         vectors = gen_vectors(nq, dim)
         table_list = []
         for i in range(10):
-            table_name = gen_unique_str('test_create_index_multi_tables')
+            table_name = gen_unique_str()
             table_list.append(table_name)
             param = {'table_name': table_name,
                      'dimension': dim,
@@ -328,7 +343,7 @@ class TestIndexBase:
     def test_describe_index_table_not_existed(self, connect):
         '''
         target: test describe index interface when table name not existed
-        method: create table and add vectors in it, create index with an random table_name
+        method: create table and add vectors in it, create index
             , make sure the table name not in index
         expected: return code not equals to 0, describe index failed
         '''
@@ -349,7 +364,7 @@ class TestIndexBase:
     def test_describe_index_not_create(self, connect, table):
         '''
         target: test describe index interface when index not created
-        method: create table and add vectors in it, create index with an random table_name
+        method: create table and add vectors in it, create index
             , make sure the table name not in index
         expected: return code not equals to 0, describe index failed
         '''
@@ -373,9 +388,9 @@ class TestIndexBase:
         method: create table and add vectors in it, create index, call drop index
         expected: return code 0, and default index param
         '''
-        index_params = get_index_params
+        index_param = get_index_params
         status, ids = connect.add_vectors(table, vectors)
-        status = connect.create_index(table, index_params)
+        status = connect.create_index(table, index_param)
         assert status.OK()
         status, result = connect.describe_index(table)
         logging.getLogger().info(result)
@@ -387,15 +402,15 @@ class TestIndexBase:
         assert result._table_name == table
         assert result._index_type == IndexType.FLAT
 
-    def test_drop_index_repeatly(self, connect, table, get_simple_index_params):
+    def test_drop_index_repeatly(self, connect, table, get_index_params):
         '''
         target: test drop index repeatly
         method: create index, call drop index, and drop again
         expected: return code 0
         '''
-        index_params = get_simple_index_params
+        index_param = get_index_params
         status, ids = connect.add_vectors(table, vectors)
-        status = connect.create_index(table, index_params)
+        status = connect.create_index(table, index_param)
         assert status.OK()
         status, result = connect.describe_index(table)
         logging.getLogger().info(result)
@@ -422,7 +437,7 @@ class TestIndexBase:
     def test_drop_index_table_not_existed(self, connect):
         '''
         target: test drop index interface when table name not existed
-        method: create table and add vectors in it, create index with an random table_name
+        method: create table and add vectors in it, create index
             , make sure the table name not in index, and then drop it
         expected: return code not equals to 0, drop index failed
         '''
@@ -446,8 +461,8 @@ class TestIndexBase:
         method: create table and add vectors in it, create index
         expected: return code not equals to 0, drop index failed
         '''
-        index_params = random.choice(gen_index_params())
-        logging.getLogger().info(index_params)
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
         status, ids = connect.add_vectors(table, vectors)
         status, result = connect.describe_index(table)
         logging.getLogger().info(result)
@@ -483,7 +498,8 @@ class TestIndexBase:
         method: create index, drop index, four times, each tme use different index_params to create index
         expected: return code 0
         '''
-        index_params = random.sample(gen_index_params(), 2)
+        nlist = 16384
+        index_params = [{"index_type": IndexType.IVFLAT, "nlist": nlist}, {"index_type": IndexType.IVF_SQ8, "nlist": nlist}]
         status, ids = connect.add_vectors(table, vectors)
         for i in range(2):
             status = connect.create_index(table, index_params[i])
@@ -514,15 +530,18 @@ class TestIndexIP:
         scope="function",
         params=gen_simple_index_params()
     )
-    def get_simple_index_params(self, request):
-        yield request.param
+    def get_simple_index_params(self, request, args):
+        if "internal" not in args:
+            if request.param["index_type"] == IndexType.IVF_SQ8H:
+                pytest.skip("sq8h not support in open source")
+        return request.param
 
     """
     ******************************************************************
       The following cases are used to test `create_index` function
     ******************************************************************
     """
-
+    @pytest.mark.level(2)
     @pytest.mark.timeout(BUILD_TIMEOUT)
     def test_create_index(self, connect, ip_table, get_index_params):
         '''
@@ -543,8 +562,10 @@ class TestIndexIP:
         method: create table and add vectors in it, check if added successfully
         expected: raise exception
         '''
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
         with pytest.raises(Exception) as e:
-            status = dis_connect.create_index(ip_table, random.choice(gen_index_params()))
+            status = dis_connect.create_index(ip_table, index_param)
 
     @pytest.mark.timeout(BUILD_TIMEOUT)
     def test_create_index_search_with_query_vectors(self, connect, ip_table, get_index_params):
@@ -557,6 +578,7 @@ class TestIndexIP:
         logging.getLogger().info(index_params)
         status, ids = connect.add_vectors(ip_table, vectors)
         status = connect.create_index(ip_table, index_params)
+        assert status.OK()
         logging.getLogger().info(connect.describe_index(ip_table))
         query_vecs = [vectors[0], vectors[1], vectors[2]]
         top_k = 5
@@ -658,17 +680,20 @@ class TestIndexIP:
         method: create table and add no vectors in it, and then create index
         expected: return code equals to 0
         '''
-        status = connect.create_index(ip_table, random.choice(gen_index_params()))
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
+        status = connect.create_index(ip_table, index_param)
         assert status.OK()
 
     @pytest.mark.timeout(BUILD_TIMEOUT)
-    def test_create_index_no_vectors_then_add_vectors(self, connect, ip_table):
+    def test_create_index_no_vectors_then_add_vectors(self, connect, ip_table, get_simple_index_params):
         '''
         target: test create index interface when there is no vectors in table, and does not affect the subsequent process
         method: create table and add no vectors in it, and then create index, add vectors in it
         expected: return code equals to 0
         '''
-        status = connect.create_index(ip_table, random.choice(gen_index_params()))
+        index_param = get_simple_index_params
+        status = connect.create_index(ip_table, index_param)
         status, ids = connect.add_vectors(ip_table, vectors)
         assert status.OK()
 
@@ -679,11 +704,11 @@ class TestIndexIP:
         method: create index after index have been built
         expected: return code success, and search ok
         '''
+        nlist = 16384
         status, ids = connect.add_vectors(ip_table, vectors)
-        index_params = random.choice(gen_index_params())
-        # index_params = get_index_params
-        status = connect.create_index(ip_table, index_params)
-        status = connect.create_index(ip_table, index_params)
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
+        status = connect.create_index(ip_table, index_param)
+        status = connect.create_index(ip_table, index_param)
         assert status.OK()
         query_vec = [vectors[0]]
         top_k = 1
@@ -698,16 +723,19 @@ class TestIndexIP:
         method: create another index with different index_params after index have been built
         expected: return code 0, and describe index result equals with the second index params
         '''
+        nlist = 16384
         status, ids = connect.add_vectors(ip_table, vectors)
-        index_params = random.sample(gen_index_params(), 2)
+        index_type_1 = IndexType.IVF_SQ8
+        index_type_2 = IndexType.IVFLAT
+        index_params = [{"index_type": index_type_1, "nlist": nlist}, {"index_type": index_type_2, "nlist": nlist}]
         logging.getLogger().info(index_params)
-        status = connect.create_index(ip_table, index_params[0])
-        status = connect.create_index(ip_table, index_params[1])
-        assert status.OK()
+        for index_param in index_params:
+            status = connect.create_index(ip_table, index_param)
+            assert status.OK()
         status, result = connect.describe_index(ip_table)
-        assert result._nlist == index_params[1]["nlist"]
+        assert result._nlist == nlist
         assert result._table_name == ip_table
-        assert result._index_type == index_params[1]["index_type"]
+        assert result._index_type == index_type_2
 
     """
     ******************************************************************
@@ -741,7 +769,7 @@ class TestIndexIP:
         vectors = gen_vectors(nq, dim)
         table_list = []
         for i in range(10):
-            table_name = gen_unique_str('test_create_index_multi_tables')
+            table_name = gen_unique_str()
             table_list.append(table_name)
             param = {'table_name': table_name,
                      'dimension': dim,
@@ -783,7 +811,7 @@ class TestIndexIP:
     def test_describe_index_not_create(self, connect, ip_table):
         '''
         target: test describe index interface when index not created
-        method: create table and add vectors in it, create index with an random table_name
+        method: create table and add vectors in it, create index
             , make sure the table name not in index
         expected: return code not equals to 0, describe index failed
         '''
@@ -850,8 +878,10 @@ class TestIndexIP:
         method: drop index, and check if drop successfully
         expected: raise exception
         '''
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVFLAT, "nlist": nlist}
         with pytest.raises(Exception) as e:
-            status = dis_connect.drop_index(ip_table, random.choice(gen_index_params()))
+            status = dis_connect.drop_index(ip_table, index_param)
 
     def test_drop_index_table_not_create(self, connect, ip_table):
         '''
@@ -859,8 +889,9 @@ class TestIndexIP:
         method: create table and add vectors in it, create index
         expected: return code not equals to 0, drop index failed
         '''
-        index_params = random.choice(gen_index_params())
-        logging.getLogger().info(index_params)
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
+        logging.getLogger().info(index_param)
         status, ids = connect.add_vectors(ip_table, vectors)
         status, result = connect.describe_index(ip_table)
         logging.getLogger().info(result)
@@ -896,7 +927,8 @@ class TestIndexIP:
         method: create index, drop index, four times, each tme use different index_params to create index
         expected: return code 0
         '''
-        index_params = random.sample(gen_index_params(), 2)
+        nlist = 16384
+        index_params = [{"index_type": IndexType.IVFLAT, "nlist": nlist}, {"index_type": IndexType.IVF_SQ8, "nlist": nlist}]
         status, ids = connect.add_vectors(ip_table, vectors)
         for i in range(2):
             status = connect.create_index(ip_table, index_params[i])
@@ -927,19 +959,21 @@ class TestIndexTableInvalid(object):
     def get_table_name(self, request):
         yield request.param
 
-    # @pytest.mark.level(1)
+    @pytest.mark.level(2)
     def test_create_index_with_invalid_tablename(self, connect, get_table_name):
         table_name = get_table_name
-        status = connect.create_index(table_name, random.choice(gen_index_params()))
+        nlist = 16384
+        index_param = {"index_type": IndexType.IVF_SQ8, "nlist": nlist}
+        status = connect.create_index(table_name, index_param)
         assert not status.OK()
 
-    # @pytest.mark.level(1)
+    @pytest.mark.level(2)
     def test_describe_index_with_invalid_tablename(self, connect, get_table_name):
         table_name = get_table_name
         status, result = connect.describe_index(table_name)
         assert not status.OK()   
 
-    # @pytest.mark.level(1)
+    @pytest.mark.level(2)
     def test_drop_index_with_invalid_tablename(self, connect, get_table_name):
         table_name = get_table_name
         status = connect.drop_index(table_name)